% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD-main.R
\name{CD.run}
\alias{CD.run}
\title{Learn structure of a discrete network}
\usage{
CD.run(indata, n_levels, eor = NULL, weights = NULL, fmlam = 0.1,
  nlam = 30, eps = 1e-04, convLb = 0.01, qtol = 1e-04, gamma = 1,
  upperbound = 100)
}
\arguments{
\item{indata}{A sparsebnData object}

\item{n_levels}{A vector indicating number of levels for each variable}

\item{eor}{Active set}

\item{weights}{weight matrix}
}
\value{
A matrix, consist of all the graphs along the solution path.
}
\description{
Learn structure of a discrete network
}

