% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{coef_gen}
\alias{coef_gen}
\title{coef_gen}
\usage{
coef_gen(edge_list, n_levels = NULL, FUN = NULL, flip = TRUE)
}
\arguments{
\item{edge_list}{a \code{\link[sparsebnUtils]{edgeList}} object.}

\item{n_levels, }{a list of number of levels for each node.}

\item{FUN, }{a probability distribution to generate coefficients}

\item{flip, }{a bool parameter. If true, will randomly flip the sign of coefficients.}
}
\value{
A list of coefficient matrix
}
\description{
coefficient generating function
}
