% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{discrete_model_selection}
\alias{discrete_model_selection}
\title{discrete_model_selection}
\usage{
discrete_model_selection(dag_path, datbn, alpha = 0.3)
}
\arguments{
\item{dag_path}{\code{\link{sparsebnPath}} object.}

\item{datbn}{\code{\link{sparsebnData}} containing the original data.}

\item{alpha}{tuning parameter for selection between 0 and 1, default value is 0.3.}
}
\value{
A scaler, indicating which lambda is chosen
}
\description{
Choose the best DAG model according to the criterion described in \href{https://arxiv.org/pdf/1403.2310.pdf}{Gu, Fu and Zhou (2016)}
(Section 3.4).
}
