% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{fit_dag}
\alias{fit_dag}
\title{A function to do inference in Bayesian network.}
\usage{
fit_dag(parents, n_levels, dat)
}
\arguments{
\item{parents}{An edgeList, sparsebnFit object or adjacency matrix.}

\item{n_levels}{A vector indicating number of levels for each variable}

\item{dat}{Data, a dataframe or matrix}
}
\description{
A function to do inference in Bayesian network.
}

