% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{generate.lambdas}
\alias{generate.lambdas}
\title{Function to generate lambda sequence}
\usage{
generate.lambdas(lambda.max, lambdas.ratio = 0.01, lambdas.length = 30)
}
\arguments{
\item{lambda.max}{Maximum lambda.}

\item{lambdas.ratio}{Ratio of minimum lambda over maximum lambda.}

\item{lambdas.length}{Length of the lambda sequence.}
}
\description{
Function to generate lambda sequence
}

