% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{generate_discrete_data}
\alias{generate_discrete_data}
\title{generate_discrete_data}
\usage{
generate_discrete_data(edge_list, n, ivn = NULL, n_levels = NULL,
  coef = NULL, FUN = NULL, flip = TRUE)
}
\arguments{
\item{edge_list}{a \code{\link[sparsebnUtils]{edgeList}} object.}

\item{n}{size of the data set, a scaler}

\item{ivn, }{a list of intervention for each data point.}

\item{n_levels, }{a list of number of levels for each node, default is binary data set.}

\item{coef, }{coefficient list (optional).}

\item{FUN, }{a function to generate magnitude of influence (optional).}

\item{flip, }{a bool parameter. If true, when generating coefficients, will randomly flip the sign of coefficients.}
}
\value{
data matrix
}
\description{
A function that generate discrete data set.
}

