% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-adapt-weights.R
\name{get.adaptive.weights}
\alias{get.adaptive.weights}
\title{get.adaptive.weights}
\usage{
get.adaptive.weights(DAG, databn, error.tol = 1e-04, convLb = 0.01)
}
\arguments{
\item{DAG}{A \code{\link[sparsebnUtils]{sparsebnFit}} object.}

\item{databn}{A \code{\link[sparsebnUtils]{sparsebnData}} object.}

\item{error.tol}{Error tolerance for the algorithm, used to test for convergence.}

\item{convLb}{Small positive number used in Hessian approximation.}
}
\value{
A list of the adaptive weight matrix, and log-likelihood of estimation.
}
\description{
A function that will return adaptive weights, and log-likelihood of the estimation.
}
\details{
This function takes a dag as an input
}

