% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD-main.R
\name{cd.run}
\alias{cd.run}
\title{Structure learning of discrete Bayesian network}
\usage{
cd.run(indata, weights = NULL, lambdas = NULL, lambdas.length = 30,
  error.tol = 1e-04, convLb = 0.01, weight.scale = 1, upperbound = 100)
}
\arguments{
\item{indata}{A sparsebnData object.}

\item{weights}{Weight matrix.}

\item{lambdas}{Numeric vector containing a grid of lambda values (i.e. regularization parameters) to use in the solution path. If missing, a default grid of values will be used based on a decreasing log-scale.}

\item{lambdas.length}{Integer number of values to include in the solution path.}

\item{error.tol}{Error tolerance for the algorithm, used to test for convergence.}

\item{convLb}{Small positive number used in Hessian approximation.}

\item{weight.scale}{A postitive number to scale weight matrix.}

\item{upperbound}{A large positive value used to truncate the adaptive weights. A -1 value indicates that there is no truncation.}
}
\value{
A \code{\link[sparsebnUtils]{sparsebnPath}} object.
}
\description{
Estimate structure of a discrete Bayesian network from observational/interventional data using the CD algorithm described in \href{http://arxiv.org/abs/1403.2310}{Gu et al. (2016)}
}

