% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{data_gen}
\usage{
data_gen(edge_list, data_size, ivn = NULL, n_levels = NULL,
  magnitude = NULL)
}
\arguments{
\item{edge_list}{a \code{\link[sparsebnUtils]{edgeList}} object.}

\item{data_size}{size of the data set, a scaler}

\item{ivn, }{a list of intervention for each data point.}

\item{n_levels, }{a list of number of levels for each node.}

\item{magnitude, }{a function to generate magnitude of influence}
}
\value{
data matrix
}
\description{
data generating function
}

