% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{generate_discrete_data}
\alias{generate_discrete_data}
\title{generate_discrete_data}
\usage{
generate_discrete_data(graph, params, n, ivn = NULL, n_levels = NULL)
}
\arguments{
\item{graph}{a \code{\link[sparsebnUtils]{edgeList}} object.}

\item{params, }{coefficient list.}

\item{n}{size of the data set, a scaler}

\item{ivn, }{a list of intervention for each data point.}

\item{n_levels, }{a vector of number of levels for each node.}
}
\value{
data matrix
}
\description{
data generating function
}
